Fix various -Wformat problems.
Updating gnulib has caused -Wformat-signedness to be enabled.  This
has revealed many problems in C format strings.  The fixes here fall
into the following main categories:

 - Using %d with an unsigned parameter.

 - %x and %o expect an unsigned argument.

 - uid_t and gid_t are unsigned on Linux.  The safe way to print these
   is to cast them to uintmax_t and then print them using the %ju
   modifier (see http://stackoverflow.com/a/1401581).

 - Using %d to print an enum.  Since enums may be either char or int,
   I fixed this by casting the enum to int.

 - strtol_error & lzma_ret are both unsigned types.

(cherry picked from commit ccdbbc7fe43e313c16a00dd30c57a04134e60b2c)

+++ cat/ls.c
@@ -699,7 +699,7 @@ output_int64_perms (int64_t i)
 {
   next_field ();
   /* csv doesn't need escaping */
-  if (printf ("%04" PRIo64, i) < 0) {
+  if (printf ("%04" PRIo64, (uint64_t) i) < 0) {
     perror ("printf");
     exit (EXIT_FAILURE);
   }
@@ -771,7 +771,8 @@ output_int64_dev (int64_t i)
   next_field ();
 
   /* csv doesn't need escaping */
-  if (printf ("%d:%d", major (dev), minor (dev)) < 0) {
+  if (printf ("%ju:%ju",
+              (uintmax_t) major (dev), (uintmax_t) minor (dev)) < 0) {
     perror ("printf");
     exit (EXIT_FAILURE);
   }
